# xcd remapping utils and fixed accuracy issue when group_size_m = 8 persistent MoE