- Made XCD remapping utils so future kernels can use them
- Added XCD remapping to persistent MoE
- Refined the pid logic in the persistent MoE so fixed accuracy issue when group_size_m = 8, and we don't need the while loop

### Old PID Logic (without XCD remapping)

| Model        |   M   |    N   |   K   | E | Top-K | Time (ms) |   TFLOPS   | Bandwidth (GB/s) |
|--------------|-------|--------|-------|---|--------|-----------|------------|------------------|
| mistral-7B   | 4096  | 14336  | 4096  | 8 |   2    | 1.890768  | 435.816124 | 630.139166       |
| mistral-7B   | 4096  | 4096   | 7168  | 8 |   2    | 0.883483  | 542.689345 | 664.241009       |
| mistral-22B  | 4096  | 16384  | 6144  | 8 |   2    | 3.794656  | 510.266333 | 645.857195       |
| mistral-22B  | 4096  | 6144   | 8192  | 8 |   2    | 1.531548  | 535.027298 | 633.237635       |
| deepseek-V3  | 4096  | 18432  | 7168  | 8 |   2    | 4.401492  | 499.536502 | 563.072025       |
| deepseek-V3  | 4096  | 7168   | 9216  | 8 |   2    | 2.248900  | 471.025455 | 540.548309       |

### New PID Logic (without XCD remapping)

| Model        |   M   |    N   |   K   | E | Top-K | Time (ms) |   TFLOPS   | Bandwidth (GB/s) |
|--------------|-------|--------|-------|---|--------|-----------|------------|------------------|
| mistral-7B   | 4096  | 14336  | 4096  | 8 |   2    | 1.863226  | 524.698037 | 621.526329       |
| mistral-7B   | 4096  | 4096   | 7168  | 8 |   2    | 0.892475  | 537.959916 | 672.667479       |
| mistral-22B  | 4096  | 16384  | 6144  | 8 |   2    | 2.982181  | 549.903978 | 645.305769       |
| mistral-22B  | 4096  | 6144   | 8192  | 8 |   2    | 1.523846  | 536.469422 | 631.047862       |
| deepseek-V3  | 4096  | 18432  | 7168  | 8 |   2    | 4.221931  | 506.613343 | 569.864464       |
| deepseek-V3  | 4096  | 7168   | 9216  | 8 |   2    | 2.612201  | 456.987797 | 554.731439       |

### New PID Logic (with XCD remapping)

| Model        |   M   |    N   |   K   | E | Top-K | Time (ms) |   TFLOPS   | Bandwidth (GB/s) |
|--------------|-------|--------|-------|---|--------|-----------|------------|------------------|
| mistral-7B   | 4096  | 14336  | 4096  | 8 |   2    | 1.732377  | 501.942072 | 681.071600       |
| mistral-7B   | 4096  | 4096   | 7168  | 8 |   2    | 0.882937  | 542.483522 | 671.668137       |
| mistral-22B  | 4096  | 16384  | 6144  | 8 |   2    | 2.797813  | 587.038480 | 620.911332       |
| mistral-22B  | 4096  | 6144   | 8192  | 8 |   2    | 1.497582  | 552.048974 | 645.617094       |
| deepseek-V3  | 4096  | 18432  | 7168  | 8 |   2    | 3.798243  | 576.953408 | 669.087333       |
| deepseek-V3  | 4096  | 7168   | 9216  | 8 |   2    | 2.200882  | 488.069659 | 568.121355